pyproject.toml
[tool.poetry]
name = "kafka"
version = "0.1.0"
description = ""
authors = ["AyumiOsawa"]
packages = [
  { include = "app" },
] 

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"

[tool.poetry.scripts]
enqueue_messages_cli = 'app.producer:main'
enqueue_json_cli = 'app.enqueue_json:main'
dequeue_messages_cli = 'app.consumer:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
